+++ rog-platform/src/hid_raw.rs
@@ -44,7 +44,10 @@ impl HidRaw {
                 let dev_path = device.devpath().to_string_lossy();
                 if dev_path.contains("virtual") && dev_path.contains(&id_product.to_uppercase()) {
                     if let Some(dev_node) = device.devnode() {
-                        info!("Using device at: {:?} for asdfgsadfgh control", dev_node);
+                        info!(
+                            "Using device at: {:?} for <TODO: label control> control",
+                            dev_node
+                        );
                         return Ok(Self(dev_node.to_owned()));
                     }
                 }